feat(serve): Qwen3.8-Flash-Next serves with its NextN draft head
--spec mtp now runs the Flash-Next MTP head end to end. One 5090, native GGUF
read in place, 80-token completions:

  off              30.6 tok/s
  --draft-tokens 3 34.1 tok/s   3.29 tok/round, 78.6 % accepted
  --draft-tokens 5 34.1 tok/s   4.65 tok/round, 73.8 % accepted

The acceptance rate is the evidence the graph is right: a head folded or
collapsed wrongly drafts noise and accepts nothing.

The head's block is an ordinary trunk full-attention block, so it is run by the
trunk's own mixer rather than a second program. `attn_mix` takes a KV plane so
the head can write its own; the target supplies the fold that seeds the
residual, the collapse that reads it back, and the block's weights. Targets
that say nothing keep the fixed Qwen3.5-shaped draft tail untouched, which is
what `mtp_block_is_trunk_layer` selects between.

The head folds into the model's *wide* hyper-connection residual, so the hidden
that crosses a round boundary is the residual, not the model width, whenever
one is running -- into the head, into the continuation store, out of the
verify. `TextGeometry::residual` already meant this and equals `hidden` for
every other family, so nothing else moves. Where the LM head still wants the
collapsed width it gets it from a scratch, in `finish_round_hidden`,
`finish_prefill` and `lm_head_view`.

Four things the head needed that the target had never been asked for:

- `hyper_connection_norm`: the mix's first line on its own, for the wide
  residual's stream-wise norm.
- `mtp_pack_fc_input_streams`: the fold's pack with the embedding broadcast
  across streams and the stream fastest, so one matmul writes a stream-major
  residual back.
- `ngram_ple_expand_columns`: a speculative round gives each lane several
  consecutive columns, and the PLE's n-gram history spans them; the prologue
  had only ever seen one column per lane.
- `gdn_input_projection_record`: a verify records the pre-convolution
  projection so a rejected draft replays the scan instead of rolling the
  recurrent state back. The projection is the layer's own, the convolution tail
  the family's. Its fold geometry (36 GDN layers, 16x48 heads) joins the
  registered list -- the same instantiation as the 48-layer entry, whose layer
  count is an identity tag the kernel never reads.

Also fixes the scan ring the auto-sized pool could ask for: it must be at least
one expert set and at most half the pool, so it needs two sets before it fits,
and a pool between 1.5 and 2 sets was refused by the directory.

Every other target still serves: llama, gemma3, qwen3_5 checked by hand, and
the expert-slot-cache, cpu-expert, gated-delta-net and ggml-k tests pass.

+++ surogate/serve/convert/qwen4exp/inventory.py
@@ -223,7 +223,9 @@ def _build_mtp_specs() -> tuple[TensorSpec, ...]:
     prefix = "mtp/"
     layer = prefix + "layer/"
     specs: list[TensorSpec] = [
-        tensor_spec(prefix + "embedding_norm", (HIDDEN,), FP32),
+        # BF16 because the plain RMSNorm reads BF16 gammas; the stream-wise norm the wide
+        # residual takes reads FP32, as every hyper-connection norm does.
+        tensor_spec(prefix + "embedding_norm", (HIDDEN,), BF16),
         tensor_spec(prefix + "hidden_norm", (HC_WIDTH,), FP32),
         tensor_spec(prefix + "input_projection", (HIDDEN, 2 * HIDDEN), W8),
     ]
